Move kinto scripts out of core
